Upgrade packages and move messages into src folder
