+++ package.json
@@ -15,10 +15,8 @@
     "@tanstack/react-query": "^5.55.0",
     "axios": "^1.7.7",
     "clsx": "^2.1.1",
-    "next": "^14.2.8",
-    "next-intl": "^3.19.1",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^15.1.4",
+    "next-intl": "^3.26.3",
     "react-hook-form": "^7.53.0",
     "zod": "^3.23.8"
   },
@@ -29,13 +27,13 @@
     "@types/node": "^22.5.4",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@vitejs/plugin-react": "^4.3.1",
     "eslint": "^8",
     "eslint-config-next": "^14.2.8",
-    "sass": "^1.78.0",
-    "typescript": "^5",
-    "vitest": "^2.0.5",
     "jsdom": "^25.0.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "tsx": "^4.19.0"
+    "sass": "^1.78.0",
+    "tsx": "^4.19.0",
+    "typescript": "^5",
+    "vitest": "^2.0.5"
   }
 }
