+++ src/pages/index.tsx
@@ -18,7 +18,7 @@ export async function getStaticProps(context: any) {
       // You can get the messages from anywhere you like. The recommended
       // pattern is to put them in JSON files separated by locale and read
       // the desired one based on the `locale` received from Next.js.
-      messages: (await import(`@/../messages/${context.locale}.json`)).default
+      messages: (await import(`@/messages/${context.locale}.json`)).default
     }
   };
 }
